refactor: 🚚 Adjust the code directory structure
