+++ packages/cli/src/commands/type/index.ts
