+++ packages/cli/src/commands/type/schema.d.ts
